+++ src/store.js
@@ -270,6 +270,11 @@ const [previewPadding, setPreviewPadding] = setter('previewPadding', 2);
 
 const [previewScaling, setPreviewScaling] = setter('previewScaling', [1, 1]);
 
+const [
+  previewScaleToCover,
+  setPreviewScaleToCover
+] = setter('previewScaleToCover', [false, false]);
+
 const [previewSpacing, setPreviewSpacing] = setter('previewSpacing', 2);
 
 const [previewOffset, setPreviewOffset] = setter('previewOffset', 2);
@@ -716,11 +721,12 @@ const createStore = () => {
     previewBorderColor,
     previewBorderOpacity,
     previewItemOffset,
+    previewOffset,
     previewPadding,
     previewRenderer,
+    previewScaleToCover,
     previewScaling,
     previewSpacing,
-    previewOffset,
     projector,
     rowHeight,
     splittingObjective,
@@ -888,11 +894,12 @@ export const createAction = {
   setPreviewBorderColor,
   setPreviewBorderOpacity,
   setPreviewItemOffset,
+  setPreviewOffset,
   setPreviewPadding,
   setPreviewRenderer,
+  setPreviewScaleToCover,
   setPreviewScaling,
   setPreviewSpacing,
-  setPreviewOffset,
   setProjector,
   setRowHeight,
   setSplittingObjective,